FEATURE: Move hardcoding of commands that require sudo in Web Service
FIXES - JIRA-739

Currently, there are several commands in the Stacki command line, that,
when invoked through the web-service, require sudo access. These commands
are hardcoded into the web-service.

This check-in adds the ability to dynamically update list of commands
that require sudo access.

INTERNAL: Standardize JSON output headers

There is currently no standard for Output headers.
When printing on the command line, all headers are
changed to uppercase. However, when using output-format=json
the headers are kept as they are in code. Which means, that
any consumer needs to know what the case for the headers are.

This check-in cleans it up for the stack WS API commands

BREAKING CHANGE: Any commands that currently consume the output
of `stack {verb} api` commands will need to be audited to
ensure that the headers are being referenced correctly.

INTERNAL: Clean up how we check for commands

Currently, we have a haphazard way of checking if a command
is valid or not. Consolidate this method across multiple sudo
and blacklist commands.

Also, fix error messages when for `remove api [user | group] perms`

INTERNAL: Test cases for api manipulation commands

Write test cases for some api commands

INTERNAL: Fix dump api plugin